pyproject.toml
[tool.poetry]
name = "assignment-levontin"
version = "0.1.0"
description = ""
authors = ["Yona <yonabitton@gmail.com>"]
readme = "README.md"


packages = [
  { include = "lib" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10"
streamlit = ">=1.47.1,<2.0.0"
pandas = ">=2.3.1,<3.0.0"
matplotlib = ">=3.7"
scikit-learn = "~1.3"
plotly = "^5.18"
loguru = "^0.7"
notebook = "^7.5.0"
jupyterlab = "^4.5.0"
seaborn = ">=0.13"
openai = "^2.8.1"
python-dotenv = "^1.2.1"
panel = "^1.8.3"
jupyter-bokeh = "^4.0.5"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.1.6"
ipykernel = "^6.27.1"
nbstripout = "^0.6.1"
codespell = "^2.2.6"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
sqlfluff = "^2.3.5"


[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
ignore = [
    "D100",  # Missing docstring in public module
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in `__init__`
    "D400",  # First line should end with a period
    "D415",  # First line should end with a period
    "D401",  # First line should be in imperative mood
]
select = [
    "B",    # bugbear
    "C4",   # comprehensions
    "C90",  # mccabe
    "D",    # docstrings
    "E",    # flake8
    "F",    # flake8
    "W",    # flake8
    "S",    # bandit
    "N",    # pep8-naming
    "RUF",  # ruff
    "PD",   # pandas
    "I",    # isort
    "ANN",  # annotations and typing
    "T20",  # print statements
    "ARG",  # unused arguments
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
"*__init__.py" = [
    "F401",
    "D100",
    "D101",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "C408"
]
